rename HomePage to MainPage and modify Component direcotry structure #23
