+++ src/Components/Thumbnail/index.jsx
