+++ src/Components/Thumbnail/index.scss
