+++ src/Pages/MainPage/index.scss
@@ -1,111 +1,29 @@
 /*
-  Main Page Styling
+  Home Page Styling
 */
 
-@mixin content-full {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-}
-
 .mainPage {
-  width: 100%;
-  height: 100%;
-  overflow-x: hidden;
-  overflow-y: hidden;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-content: center;
+  align-items: center;
+  margin-top: 80px;
 
-  &__screen {
-    @include content-full;
+  &__newPost {
+    width: 500px;
+    border: 1px solid #ccd0d5;
+    border-radius: 20px;
+    margin-bottom: 10px;
+    text-align: center;
+    cursor: pointer;
+    &:hover {
+      background-color: #f8f8f8;
+    }
+  }
+
+  &__feed {
     display: flex;
     flex-direction: column;
-    justify-content: center;
-    align-content: center;
-    text-align: center;
-    overflow-x: hidden;
-    overflow-y: hidden;
-
-    &__intro {
-      @include content-full;
-      height: 100%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      &__part {
-        margin-top: -5%;
-
-        &-logo {
-          font-family: 'Mansalva', cursive;
-          font-weight: 400;
-          font-style: normal;
-          font-size: 7.8rem;
-          padding: 0;
-          margin: 0;
-        }
-
-        &-text {
-          padding-top: 0.5%;
-          margin: 0;
-          font-size: 1.6rem;
-        }
-
-        &-btn {
-          padding-top: 3%;
-        }
-      }
-    }
-
-    &__signIn {
-      @include content-full;
-      height: 100%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      &-none {
-        visibility: hidden;
-      }
-
-      &__part {
-        margin-top: -3%;
-
-        &__container {
-          padding: 100px 65px 90px 65px;
-          transition: 0.5s;
-          &:hover,
-          &-on {
-            background-color: rgba(255, 255, 255, 0.6);
-            box-shadow: 3px 4px 7px 5px rgba(0, 0, 0, 0.1);
-          }
-        }
-
-        &__email {
-          margin: 15px 0;
-        }
-
-        &__password {
-          margin: 15px 0;
-        }
-
-        &-btn {
-          padding-top: 10%;
-          margin-bottom: 2%;
-        }
-
-        &__signUpLink a {
-          font-size: 1rem;
-          border-bottom: 1px solid black;
-          transition: 0.3s;
-
-          &:hover {
-            padding: 0 3px;
-            border-bottom: 1px solid $main-color;
-            color: $main-color;
-          }
-        }
-      }
-    }
   }
 }
